requirements.txt
opencv-python
face_recognition
SpeechRecognition
pyttsx3
pyaudio
openai
numpy
facenet_pytorch
torch
sounddevice
vosk
googletrans
edge_tts
transformers
pvporcupine
rapidfuzz
playsound
coqui-tts